update git workflows

+++ .tools/run_python_tests.sh
@@ -6,12 +6,21 @@ SELF_PATH=${BASH_SOURCE[0]:-"$(command -v -- "$0")"}
 PROJECT_ROOT="$(dirname "$SELF_PATH")/.."
 
 function python_mypi_lint() {
-  python3 -m venv .venv
-  source .venv/bin/activate
-  pip install -r requirements.txt
-  pip install mypy
-  python3 -m mypy .
-  deactivate
+  if [ -f "pyproject.toml" ]; then
+    uv sync
+    uv add --dev mypy
+    uv run mypy .
+  elif [ -f "requirements.txt" ]; then
+    python3 -m venv .venv
+    source .venv/bin/activate
+    pip install -r requirements.txt
+    pip install mypy
+    python3 -m mypy .
+    deactivate
+  else
+    echo "No pyproject.toml or requirements.txt found in $(pwd)"
+    exit 1
+  fi
 }
 
 pushd $PROJECT_ROOT/python/templates/python && python_mypi_lint && popd
